Leon/verify board

+++ leon-go/src/main/scala/go/core/Game.scala
@@ -15,9 +15,9 @@ case class Game(states: List[Board], steps: List[Step]) {
 
   val activePlayer: PlayerType = if (states.size.toInt % 2 == 1) BlackPlayer else WhitePlayer
 
-  val round: Int = states.size.toInt - 1
+  val round: BigInt = states.size.toInt - 1
 
-  val size: Int = state.n
+  val size: BigInt = state.n
 
   def isOver: Boolean = {
     round > 1 && steps.head == Pass && steps.tail.head == Pass
@@ -43,5 +43,5 @@ case class Game(states: List[Board], steps: List[Step]) {
 
 object Game {
   def apply(b: Board): Game = Game(List[Board](b), List[Step]())
-  def apply(n: Int): Game = Game(new Board(n))
+  def apply(n: BigInt): Game = Game(new Board(n))
 }
