+++ leon-go/src/main/scala/go/core/Step.scala
@@ -4,7 +4,7 @@ import go.core.CellObject.EmptyCell
 
 sealed abstract class Step
 case object Pass extends Step
-case class Place(x: Int, y: Int) extends Step
+case class Place(x: BigInt, y: BigInt) extends Step
 
 object Step {
   def isValid(s: Step, b: Board): Boolean = {
